[tests] merge thread_discovery.py into mesh_cop.py
