+++ tests/scripts/thread-cert/mesh_cop.py
@@ -1,29 +1,68 @@
+#!/usr/bin/env python
+#
+#  Copyright (c) 2018, The OpenThread Authors.
+#  All rights reserved.
+#
+#  Redistribution and use in source and binary forms, with or without
+#  modification, are permitted provided that the following conditions are met:
+#  1. Redistributions of source code must retain the above copyright
+#     notice, this list of conditions and the following disclaimer.
+#  2. Redistributions in binary form must reproduce the above copyright
+#     notice, this list of conditions and the following disclaimer in the
+#     documentation and/or other materials provided with the distribution.
+#  3. Neither the name of the copyright holder nor the
+#     names of its contributors may be used to endorse or promote products
+#     derived from this software without specific prior written permission.
+#
+#  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+#  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+#  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+#  ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+#  LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+#  CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+#  SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+#  INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+#  CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+#  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+#  POSSIBILITY OF SUCH DAMAGE.
+#
 
 import io
 import struct
 
 from enum import IntEnum
+from network_data import SubTlvsFactory
 
 class TlvType(IntEnum):
-    STATE = 16,
-    PROVISIONING_URL = 32,
-    VENDOR_NAME = 33,
-    VENDOR_MODEL = 34,
-    VENDOR_SW_VERSION = 35,
-    VENDOR_DATA = 36,
-    VENDOR_STACK_VERSION = 37,
+    EXTENDED_PANID = 2
+    NETWORK_NAME = 3
+    STEERING_DATA = 8
+    COMMISSIONER_UDP_PORT = 15
+    STATE = 16
+    JOINER_UDP_PORT = 18
+    PROVISIONING_URL = 32
+    VENDOR_NAME = 33
+    VENDOR_MODEL = 34
+    VENDOR_SW_VERSION = 35
+    VENDOR_DATA = 36
+    VENDOR_STACK_VERSION = 37
+    DISCOVERY_REQUEST = 128
+    DISCOVERY_RESPONSE = 129
+
 
 class MeshCopMessageType(IntEnum):
-    JOIN_FIN_REQ = 1,
-    JOIN_FIN_RSP = 2,
-    JOIN_ENT_NTF = 3,
+    JOIN_FIN_REQ = 1
+    JOIN_FIN_RSP = 2
+    JOIN_ENT_NTF = 3
     JOIN_ENT_RSP = 4
+
 
 def create_mesh_cop_message_type_set():
     return [ MeshCopMessageType.JOIN_FIN_REQ,
              MeshCopMessageType.JOIN_FIN_RSP,
              MeshCopMessageType.JOIN_ENT_NTF,
              MeshCopMessageType.JOIN_ENT_RSP ]
+
 
 class State(object):
 
@@ -47,6 +86,7 @@ class StateFactory:
         state = ord(data.read(1))
         return State(state)
 
+
 class VendorName(object):
 
     def __init__(self, vendor_name):
@@ -62,11 +102,13 @@ class VendorName(object):
     def __repr__(self):
         return "VendorName(vendor_name={})".format(self.vendor_name)
 
+
 class VendorNameFactory:
 
     def parse(self, data):
         vendor_name = data.getvalue().decode('utf-8')
         return VendorName(vendor_name)
+
 
 class VendorModel(object):
 
@@ -82,6 +124,7 @@ class VendorModel(object):
 
     def __repr__(self):
         return "VendorModel(vendor_model={})".format(self.vendor_model)
+
 
 class VendorModelFactory:
 
@@ -104,11 +147,13 @@ class VendorSWVersion(object):
     def __repr__(self):
         return "VendorName(vendor_sw_version={})".format(self.vendor_sw_version)
 
+
 class VendorSWVersionFactory:
 
     def parse(self, data):
         vendor_sw_version = data.getvalue()
         return VendorSWVersion(vendor_sw_version)
+
 
 class VendorStackVersion(object):
 
@@ -123,23 +168,26 @@ class VendorStackVersion(object):
     def __repr__(self):
         return "VendorStackVersion(vendor_stack_version={})".format(self._stack_vendor_oui)
 
+
 class VendorStackVersionFactory:
 
     def parse(self, data):
         stack_vendor_oui = struct.unpack(">H", data.read(2))[0]
         return VendorStackVersion(stack_vendor_oui)
 
+
 class ProvisioningUrl(object):
 
     def __init__(self, url):
         self._url = url
-    
+
     @property
     def url(self):
         return self._url
-    
+
     def __repr__(self):
         return "ProvisioningUrl(url={})".format(self.url)
+
 
 class ProvisioningUrlFactory:
 
@@ -147,11 +195,12 @@ class ProvisioningUrlFactory:
         url = data.decode('utf-8')
         return ProvisioningUrl(url)
 
+
 class VendorData(object):
 
     def __init__(self, data):
         self._vendor_data = data
-    
+
     @property
     def vendor_data(self):
         return self._vendor_data
@@ -159,10 +208,12 @@ class VendorData(object):
     def __repr__(self):
         return "Vendor(url={})".format(self.vendor_data)
 
+
 class VendorDataFactory(object):
 
     def parse(self, data):
         return VendorData(data)
+
 
 class MeshCopCommand(object):
 
@@ -182,6 +233,7 @@ class MeshCopCommand(object):
         tlvs_str = ", ".join(["{}".format(tlv) for tlv in self.tlvs])
         return "MeshCopCommand(type={}, tlvs=[{}])".format(self.type, tlvs_str)
 
+
 def create_deault_mesh_cop_msg_type_map():
     return {
         'JOIN_FIN.req': MeshCopMessageType.JOIN_FIN_REQ,
@@ -189,6 +241,7 @@ def create_deault_mesh_cop_msg_type_map():
         'JOIN_ENT.ntf': MeshCopMessageType.JOIN_ENT_NTF,
         'JOIN_ENT.rsp': MeshCopMessageType.JOIN_ENT_RSP
     }
+
 
 class MeshCopCommandFactory:
 
@@ -241,3 +294,149 @@ def create_default_mesh_cop_tlv_factories():
         TlvType.VENDOR_DATA: VendorDataFactory(),
         TlvType.VENDOR_STACK_VERSION: VendorStackVersionFactory()
     }
+
+
+class ThreadDiscoveryTlvsFactory(SubTlvsFactory):
+
+    def __init__(self, sub_tlvs_factories):
+        super(ThreadDiscoveryTlvsFactory, self).__init__(sub_tlvs_factories)
+
+
+class DiscoveryRequest(object):
+
+    def __init__(self, version, joiner_flag):
+        self._version = version
+        self._joiner_flag = joiner_flag
+
+    @property
+    def version(self):
+        return self._version
+
+    @property
+    def joiner_flag(self):
+        return self._joiner_flag
+
+    def __eq__(self, other):
+        return (type(self) is type(other)
+            and self.version == other.version
+            and self.joiner_flag == other.joiner_flag)
+
+    def __repr__(self):
+        return "DiscoveryRequest(version={}, joiner_flag={})".format(
+            self.version, self.joiner_flag)
+
+
+class DiscoveryRequestFactory(object):
+
+    def parse(self, data, message_info):
+        data_byte = struct.unpack(">B", data.read(1))[0]
+        version = (data_byte & 0xf0) >> 4
+        joiner_flag = (data_byte & 0x08) >> 3
+
+        return DiscoveryRequest(version, joiner_flag)
+
+
+class DiscoveryResponse(object):
+
+    def __init__(self, version, native_flag):
+        self._version = version
+        self._native_flag = native_flag
+
+    @property
+    def version(self):
+        return self._version
+
+    @property
+    def native_flag(self):
+        return self._native_flag
+
+    def __eq__(self, other):
+        return (type(self) is type(other)
+            and self.version == other.version
+            and self.native_flag == other.native_flag)
+
+    def __repr__(self):
+        return "DiscoveryResponse(version={}, native_flag={})".format(
+            self.version, self.native_flag)
+
+
+class DiscoveryResponseFactory(object):
+
+    def parse(self, data, message_info):
+        data_byte = struct.unpack(">B", data.read(1))[0]
+        version = (data_byte & 0xf0) >> 4
+        native_flag = (data_byte & 0x08) >> 3
+
+        return DiscoveryResponse(version, native_flag)
+
+
+class ExtendedPanid(object):
+
+    def __init__(self, extended_panid):
+        self._extended_panid = extended_panid
+
+    @property
+    def extended_panid(self):
+        return self._extended_panid
+
+    def __eq__(self, other):
+        return (type(self) is type(other)
+            and self.extended_panid == other.extended_panid)
+
+    def __repr__(self):
+        return "ExtendedPanid(extended_panid={})".format(self.extended_panid)
+
+
+class ExtendedPanidFactory(object):
+
+    def parse(self, data, message_info):
+        extended_panid = struct.unpack(">Q", data.read(8))[0]
+        return ExtendedPanid(extended_panid)
+
+
+class NetworkName(object):
+
+    def __init__(self, network_name):
+        self._network_name = network_name
+
+    @property
+    def network_name(self):
+        return self._network_name
+
+    def __eq__(self, other):
+        return (type(self) is type(other)
+            and self.network_name == other.network_name)
+
+    def __repr__(self):
+        return "NetworkName(network_name={})".format(self.network_name)
+
+
+class NetworkNameFactory(object):
+
+    def parse(self, data, message_info):
+        len = message_info.length
+        network_name = struct.unpack("{}s".format(10), data.read(len))[0]
+        return NetworkName(network_name)
+
+
+class JoinerUdpPort(object):
+
+    def __init__(self, udp_port):
+        self._udp_port = udp_port
+
+    @property
+    def udp_port(self):
+        return self._udp_port
+
+    def __eq__(self, other):
+        return type(self) is type(other) and self.udp_port == other.udp_port
+
+    def __repr__(self):
+        return "JoinerUdpPort(udp_port={})".format(self.udp_port)
+
+
+class JoinerUdpPortFactory(object):
+
+    def parse(self, data, message_info):
+        udp_port = struct.unpack(">H", data.read(2))[0]
+        return JoinerUdpPort(udp_port)
